requirements.txt
discord.py==2.4.0
Flask==3.0.3
openpyxl==3.1.5
filelock==3.16.1
Flask-Cors==5.0.0
Gunicorn